+++ project/clip.h
@@ -33,17 +33,17 @@ class QOpenGLTexture;
 
 struct Clip
 {
-    Clip(Sequence* s);
+	Clip(Sequence* s);
 	~Clip();
-    Clip* copy(Sequence* s);
-    void reset_audio();
+	Clip* copy(Sequence* s);
+	void reset_audio();
 	void reset();
 	void refresh();
-    long get_clip_in_with_transition();
+	long get_clip_in_with_transition();
 	long get_timeline_in_with_transition();
 	long get_timeline_out_with_transition();
 	long getLength();
-    double getMediaFrameRate();
+	double getMediaFrameRate();
 	long getMaximumLength();
 	void recalculateMaxLength();
 	int getWidth();
@@ -56,39 +56,39 @@ struct Clip
 	void queue_remove_earliest();
 
 	// timeline variables (should be copied in copy())
-    bool enabled;
-    long clip_in;
-    long timeline_in;
-    long timeline_out;
-    int track;
+	bool enabled;
+	long clip_in;
+	long timeline_in;
+	long timeline_out;
+	int track;
 	QString name;
-    quint8 color_r;
-    quint8 color_g;
-    quint8 color_b;    
-    Media* media;
-    int media_stream;	
+	quint8 color_r;
+	quint8 color_g;
+	quint8 color_b;
+	Media* media;
+	int media_stream;
 	double speed;
-    double cached_fr;
+	double cached_fr;
 	bool reverse;
 	bool maintain_audio_pitch;
 	bool autoscale;
 
 	// other variables (should be deep copied/duplicated in copy())
-    QList<Effect*> effects;
-    QVector<int> linked;
-    int opening_transition;
-    Transition* get_opening_transition();
-    int closing_transition;
-    Transition* get_closing_transition();
+	QList<Effect*> effects;
+	QVector<int> linked;
+	int opening_transition;
+	Transition* get_opening_transition();
+	int closing_transition;
+	Transition* get_closing_transition();
 
 	// media handling
-    AVFormatContext* formatCtx;
-    AVStream* stream;
-    AVCodec* codec;
-    AVCodecContext* codecCtx;
-    AVPacket* pkt;
+	AVFormatContext* formatCtx;
+	AVStream* stream;
+	AVCodec* codec;
+	AVCodecContext* codecCtx;
+	AVPacket* pkt;
 	AVFrame* frame;
-    AVDictionary* opts;
+	AVDictionary* opts;
 	long calculated_length;
 
 	// temporary variables
@@ -96,23 +96,23 @@ struct Clip
 	bool undeletable;
 	bool reached_end;
 	bool pkt_written;
-    bool open;
-    bool finished_opening;
-    bool replaced;
+	bool open;
+	bool finished_opening;
+	bool replaced;
 	bool ignore_reverse;
 	int pix_fmt;
 
 	// caching functions
 	bool use_existing_frame;
-    bool multithreaded;
+	bool multithreaded;
 	Cacher* cacher;
-    QWaitCondition can_cache;
+	QWaitCondition can_cache;
 	int max_queue_size;
 	QVector<AVFrame*> queue;
 	QMutex queue_lock;
-    QMutex lock;
+	QMutex lock;
 	QMutex open_lock;
-    int64_t last_invalid_ts;
+	int64_t last_invalid_ts;
 
 	// converters/filters
 	AVFilterGraph* filter_graph;
@@ -121,15 +121,15 @@ struct Clip
 
 	// video playback variables
 	QOpenGLFramebufferObject** fbo;
-    QOpenGLTexture* texture;
+	QOpenGLTexture* texture;
 	long texture_frame;
 
 	// audio playback variables
 	int64_t reverse_target;
-    int frame_sample_index;
-    int audio_buffer_write;
-    bool audio_reset;
-    bool audio_just_reset;
+	int frame_sample_index;
+	unsigned long audio_buffer_write;
+	bool audio_reset;
+	bool audio_just_reset;
 	long audio_target_frame;
 };
